Added iOS section, fixed double "for"
Added section on Microsoft Defender for Endpoint on iOS now that iOS is in public preview and soon to be GA. Also updated each section titles to remove the second "for" in the title after "for Endpoint". Did the same in the first sentence of each section. Changed "for" to "on" ex: Microsoft Defender for Endpoint on macOS.

Updated "Mac" to "macOS" to refer to the OS name and not the device name.